oz_be_django/pyproject.toml
[tool.poetry]
name = "oz-be-django"
version = "0.1.0"
description = ""
authors = ["박준혁 <jun201908@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
